third_party/WebKit/LayoutTests/bluetooth/requestDevice/name-missing-device-from-service-filter.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return setBluetoothFakeAdapter('NoNameHeartRateAdapter')
  .then(() => requestDeviceWithKeyDown({
    filters: [{services: ['heart_rate']}]}))
  .then(device => {
      assert_equals(device.name, null);
  })
}, 'An unnamed device can be obtained by advertised service UUID.');
</script>
